Structure and relocate Main page and Role comopnent
